requirements.txt
alembic==0.9.7
aniso8601==2.0.0
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
six==1.11.0
SQLAlchemy==1.2.2
Werkzeug==0.14.1
python-dotenv==0.14.0